+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **June 01, 2023**
 - Frontend
   - Added checkmark renderer for table component
+  - Added health indicator renderer for table component
 
 - Removed fastify-markdown plugin (DESCW-1132)
 - Backend